Add Daemon Version to each Peer entry
In this commit I added support to keep track of the daemon version
of each peer entry. I made the handshakes between each daemon
compatible with new and old daemons, and I added an RPC method to
retrieve a daemon's white and gray peer lists in order to survey
what remote nodes the local node has made contact with.